Remove `staking.debug.genesis_state` flag
The debug genesis state for staking has been moved to
`staking/debug_state.go` and is provided in the test
node genesis provider in `tendermint/genesis_testnode.go`,
which is instantiated only in unit tests.
